pfm/drive_to_vision/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-vision
python-dotenv
functions-framework